Fix: Disallow inconsistent shareable fields
This commit adds a check during type expansion to ensure that if a
shareable field on an interface is not resolvable by all object types in
all subgraphs, it violates the `@shareable` contract. This prevents the
query planner from generating invalid query plans.

This resolves an issue where the query planner could generate plans that
fetched data from multiple subgraphs for a single field, which is not
allowed by the `@shareable` directive.

Also updated test cases for interface tests.