Handle empty responses from subgraphs and failed entity calls (#500)
This PR tackles an issue where the subgraph error could end up being
omitted entirely in the client response.
It now also handles cases where the upstream response is empty.

In cases where a subgraph call fails, there's no point to clone the
error for every entity representation. It'd result in a bloated response
with the same thing repeated multiple times (imagine fetching 1000s
representations).
Instead of cloning and adding `path` property to associate the error
with a point in GraphQL document, we attach `affectedPath` that
represents the Flatten(path) where the error occurred.