refactor(error-handling): add context to `PlanExecutionError` (#513)
This commit adds context to the `PlanExecutionError` enum. It includes
the subgraph name and the affected path in case of a projection or
header propagation failure.

It also adds a `ResultExt` trait to map `ProjectionError` and
`HeaderRuleRuntimeError` to `PlanExecutionError`.

I will improve names soon, it's a draft anyway, I just wanted to
showcase how we could do errors, instead of manually creating GraphQL
Error with code and stuff in extensions.